requirements.txt
flask-sqlalchemy
openai==0.28
python-dotenv